requirements.txt
# Core
pandas>=2.0
numpy>=1.24
pyarrow>=12.0

# Web / Scraping
requests>=2.28
beautifulsoup4>=4.11
lxml>=4.9

# NLP
nltk>=3.8
vaderSentiment>=3.3
spacy>=3.5
scikit-learn>=1.2
sentence-transformers>=2.2

# Vector DB
faiss-cpu>=1.7
chromadb>=0.4

# LLM (free tier, optional)
google-generativeai>=0.3

# Dashboard
streamlit>=1.28
plotly>=5.15

# Data
yfinance>=0.2
python-dotenv>=1.0

# Utilities
tqdm>=4.65
joblib>=1.3
